Alles werkt, alleen moving algoritme moet nog geïmplementeerd worden!

+++ OGP1516-Hillbillies/src/hillbillies/model/Boulder.java
@@ -7,7 +7,7 @@ package hillbillies.model;
 
 public class Boulder extends RawMaterial {
 	
-	public Boulder(double[] position, World world) {
+	public Boulder(int[] position, World world) {
 		super(position, world);
 	}
